pyproject.toml
[project]
name = "outlook-email-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "httpx>=0.28.1",
    "pytest-playwright>=0.7.1",
]
